Add Slide class to manage pages
In the current implementation, Canvas class manages pages in a
slide. I want to move the feature to the Slide class.

But there is a problem. I used "slide" for "page" in a slide. :<
Rabbit::Element::Slide should rename to Rabbit::Element::Page.

I need to rename it before I use this class in Canvas class. Because
it is confused. It breaks backward compatibility of theme API... I
need more work for theme...