Add config 'source_code_uri'
To refer to other presentation, it is useful to possible to read the source
code of the presentation.
In a slide published on rubygems.org that has a specification "homepage"
but it points uri of slide.rabbit-shocker.org.
Therefore, allow set metadata "source_code_uri" through config.yaml to
easy to access the source code of the presentation.